Bring in results of parsing the input
I struggled a bit with the architecture, where I wanted to put a struct
in a signal that depended on the lifetime of a different signal.

This obviously wasn't going to work, but after discussing it with a
special someone, she pointed out that as I was displaying as a string,
I didn't actually need the struct in the signal.